fix(render): preserve cached spread slots
