+++ src/app/tests/runtime_worker.rs
@@ -159,42 +159,6 @@ impl PdfBackend for PageDimensionFailingPdf {
     }
 }
 
-struct TwoPageRuntimePdf;
-
-impl PdfBackend for TwoPageRuntimePdf {
-    fn path(&self) -> &std::path::Path {
-        std::path::Path::new("two-page-runtime.pdf")
-    }
-
-    fn doc_id(&self) -> u64 {
-        11
-    }
-
-    fn page_count(&self) -> usize {
-        2
-    }
-
-    fn page_dimensions(&self, _page: usize) -> AppResult<(f32, f32)> {
-        Ok((100.0, 100.0))
-    }
-
-    fn render_page(&self, _page: usize, _scale: f32) -> AppResult<RgbaFrame> {
-        Err(AppError::unsupported("not needed in runtime test"))
-    }
-
-    fn extract_text(&self, _page: usize) -> AppResult<String> {
-        Err(AppError::unsupported("not needed in runtime test"))
-    }
-
-    fn extract_positioned_text(&self, _page: usize) -> AppResult<crate::backend::TextPage> {
-        Err(AppError::unsupported("not needed in runtime test"))
-    }
-
-    fn extract_outline(&self) -> AppResult<Vec<OutlineNode>> {
-        Err(AppError::unsupported("not needed in runtime test"))
-    }
-}
-
 #[test]
 fn schedule_navigation_updates_queue_and_cancellation_metrics() {
     let file = unique_temp_path("runtime_schedule.pdf");
@@ -340,172 +304,6 @@ fn prepare_current_page_uses_zero_overlay_stamp_when_decoration_falls_back() {
         .expect("prepare should succeed");
 
     assert_eq!(presenter.last_prepare_overlay_stamp, Some(0));
-}
-
-#[test]
-fn spread_canvas_slots_crop_from_shared_pan_coordinate_space() {
-    let doc = TwoPageRuntimePdf;
-    let mut runtime = RenderRuntime::default();
-    let mut presenter = TestPresenter::default();
-    for page in 0..2 {
-        runtime.l1_cache.insert(
-            RenderedPageKey::new(doc.doc_id(), page, 1.0),
-            RgbaFrame {
-                width: 100,
-                height: 50,
-                pixels: vec![page as u8; 100 * 50 * 4].into(),
-            },
-            false,
-        );
-    }
-    let mut pan = PanOffset {
-        cells_x: 8,
-        cells_y: 0,
-    };
-
-    let areas = runtime
-        .try_prepare_spread_canvas_slots_from_cache(
-            &doc,
-            &mut presenter,
-            Viewport {
-                x: 0,
-                y: 0,
-                width: 10,
-                height: 5,
-            },
-            crate::app::VisiblePageSlots {
-                anchor_page: 0,
-                trailing_page: Some(1),
-                left_page: Some(0),
-                right_page: Some(1),
-            },
-            1.0,
-            &mut pan,
-            Some((10, 10)),
-            &HighlightOverlaySnapshot::default(),
-            1,
-            20,
-        )
-        .expect("spread canvas prepare should pass")
-        .expect("cached spread should prepare");
-
-    assert_eq!(
-        areas,
-        [Some(Rect::new(0, 0, 2, 5)), Some(Rect::new(4, 0, 6, 5))]
-    );
-    assert_eq!(presenter.prepared_frame_sizes, vec![(20, 50), (60, 50)]);
-    assert_eq!(presenter.prepared_viewports.len(), 2);
-}
-
-#[test]
-fn spread_canvas_slots_keep_slot_identity_when_left_page_is_offscreen() {
-    let doc = TwoPageRuntimePdf;
-    let mut runtime = RenderRuntime::default();
-    let mut presenter = TestPresenter::default();
-    for page in 0..2 {
-        runtime.l1_cache.insert(
-            RenderedPageKey::new(doc.doc_id(), page, 1.0),
-            RgbaFrame {
-                width: 100,
-                height: 50,
-                pixels: vec![page as u8; 100 * 50 * 4].into(),
-            },
-            false,
-        );
-    }
-    let mut pan = PanOffset {
-        cells_x: 12,
-        cells_y: 0,
-    };
-
-    let areas = runtime
-        .try_prepare_spread_canvas_slots_from_cache(
-            &doc,
-            &mut presenter,
-            Viewport {
-                x: 0,
-                y: 0,
-                width: 10,
-                height: 5,
-            },
-            crate::app::VisiblePageSlots {
-                anchor_page: 0,
-                trailing_page: Some(1),
-                left_page: Some(0),
-                right_page: Some(1),
-            },
-            1.0,
-            &mut pan,
-            Some((10, 10)),
-            &HighlightOverlaySnapshot::default(),
-            1,
-            20,
-        )
-        .expect("spread canvas prepare should pass")
-        .expect("cached spread should prepare");
-
-    assert_eq!(areas, [None, Some(Rect::new(0, 0, 10, 5))]);
-    assert_eq!(presenter.prepared_slot_pages, vec![None, Some(1)]);
-}
-
-#[test]
-fn page_slots_clamp_negative_pan_before_saving_state() {
-    let doc = TwoPageRuntimePdf;
-    let mut runtime = RenderRuntime::default();
-    let mut presenter = TestPresenter::default();
-    for page in 0..2 {
-        runtime.l1_cache.insert(
-            RenderedPageKey::new(doc.doc_id(), page, 1.0),
-            RgbaFrame {
-                width: 100,
-                height: 100,
-                pixels: vec![page as u8; 100 * 100 * 4].into(),
-            },
-            false,
-        );
-    }
-    let mut pan = PanOffset {
-        cells_x: -3,
-        cells_y: -7,
-    };
-    let page_slots = [
-        (
-            Some(0),
-            Viewport {
-                x: 0,
-                y: 0,
-                width: 5,
-                height: 5,
-            },
-        ),
-        (
-            Some(1),
-            Viewport {
-                x: 5,
-                y: 0,
-                width: 5,
-                height: 5,
-            },
-        ),
-    ];
-
-    let prepared = runtime
-        .try_prepare_page_slots_from_cache(
-            &doc,
-            &mut presenter,
-            &page_slots,
-            1.0,
-            &mut pan,
-            Some((10, 10)),
-            true,
-            &HighlightOverlaySnapshot::default(),
-            1,
-        )
-        .expect("page slot prepare should pass");
-
-    assert!(prepared);
-    assert_eq!(pan, PanOffset::default());
-    assert_eq!(presenter.prepared_slot_pages, vec![Some(0), Some(1)]);
 }
 
 #[test]
